Convert useFocusOnHide to TS
